update GUI to use the design that scientific and technical justification texts are inserted into a template latex file, remove functionality to switch justification formats, use latex only (potential future work could reintroduce e.g., RST)
