--- src/main/webui/src/ProposalEditorView/justifications/edit.modal.tsx
@@ -1,55 +1,0 @@
-import {ReactElement} from "react";
-import ViewEditButton from "src/commonButtons/viewEdit.tsx";
-import {Modal} from "@mantine/core";
-import {useDisclosure, useMediaQuery} from "@mantine/hooks";
-import {JustificationProps} from "./justifications.table.tsx";
-import JustificationLatex from "./justifications.latex.tsx";
-
-
-function capitalizeFirstChar(string : string) : string {
-    return string.charAt(0).toUpperCase() + string.slice(1);
-}
-
-export default function JustificationsEditModal(justificationProps : JustificationProps)
-    : ReactElement {
-
-    const isMobile = useMediaQuery('(max-width: 75em)');
-
-    const EditButton = () : ReactElement => {
-        return (
-            <ViewEditButton
-                toolTipLabel={"edit " + justificationProps.which + " justification"}
-                onClick={open}
-            />
-        )
-    }
-
-    const ModalHtml = () : ReactElement => {
-        return (
-            <Modal
-                opened={opened}
-                onClose={props.closeModal}
-                title={capitalizeFirstChar(props.which) + " Justification"}
-                fullScreen={isMobile}
-                size="75%"
-                closeOnClickOutside={false}
-            >
-                <JustificationLatex {...props} />
-            </Modal>
-        )
-    }
-
-    const [opened, {close, open}] = useDisclosure();
-    const props = {
-        ...justificationProps,
-        closeModal: () =>{
-            close();
-        }}
-
-    return (
-        <>
-            <EditButton/>
-            <ModalHtml/>
-        </>
-    )
-}
